Added project for parsing C# code with Jison.
This initial project parses a C# source file which contains a sequence of using statements and returns an array with all the namespaces used in that source file.

Future features will be added such as parsing a class inside a source file and extracting its members (fields, properties and methods). Similar could be done for structs and interfaces.